foto export: automatisch archiveren van verwijderde foto's

+++ src/main/java/nl/spelberg/brandweer/model/impl/FileOperationsImpl.java
@@ -20,7 +20,39 @@ public class FileOperationsImpl implements FileOperations {
 
     @Override
     public void createDirectoryRecursive(String dirPath) {
-        createDirectoryRecursive(new File(dirPath));
+        File dirFile = new File(dirPath).getAbsoluteFile();
+        createDirectoryRecursive(dirFile);
+    }
+
+    @Override
+    public void deleteFile(String fileName) {
+        File file = new File(fileName);
+        boolean isDeleted = file.delete();
+        if (!isDeleted) {
+            throw new FileOperationException("Failed to delete file: " + file.getAbsolutePath());
+        }
+    }
+
+    @Override
+    public void moveFileToDir(String filePath, String toDirPath) {
+        File file = new File(filePath);
+        if (file.isFile()) {
+            File toDir = new File(toDirPath);
+            createDirectoryRecursive(toDir);
+            File toFile = new File(toDir, file.getName());
+            boolean isMoved = file.renameTo(toFile);
+            if (!isMoved) {
+                throw new FileOperationException("Failed to move file '" + file.getAbsolutePath() + "' to '" + toFile.getAbsolutePath());
+            }
+        } else {
+            throw new NotFoundFileOperationException("File '" + filePath + "' not found, or is not a normal file.");
+        }
+    }
+
+    @Override
+    public boolean fileExists(String filePath) {
+        File file = new File(filePath);
+        return file.isFile();
     }
 
     @SuppressWarnings("ResultOfMethodCallIgnored")
@@ -35,10 +67,15 @@ public class FileOperationsImpl implements FileOperations {
     @Override
     public void copyFile(String fromPath, String toPath, FileOperations.CopyOption... copyOptions) {
         try {
+            File fromFile = new File(fromPath).getAbsoluteFile();
+            if (!fromFile.isFile()) {
+                throw new NotFoundFileOperationException(fromFile);
+            }
+
             File toFile = new File(toPath);
             if (copyFileNeeded(toFile, copyOptions)) {
-                log.debug("Copying file from '" + fromPath + "' to '" + toFile.getAbsolutePath() + "' ...");
-                FileInputStream fis = new FileInputStream(fromPath);
+                log.debug("Copying file from '" + fromFile.getAbsolutePath() + "' to '" + toFile.getAbsolutePath() + "' ...");
+                FileInputStream fis = new FileInputStream(fromFile);
                 try {
                     toFile.getParentFile().mkdirs();
                     FileOutputStream fos = new FileOutputStream(toFile);
@@ -93,14 +130,13 @@ public class FileOperationsImpl implements FileOperations {
     }
 
     @Override
-    public void write(String fileName, String expectedCsv) {
+    public void write(String fileName, String s) {
         try {
             File file = new File(fileName);
-            //noinspection ResultOfMethodCallIgnored
-            file.getParentFile().mkdirs();
+            createDirectoryRecursive(file.getParentFile());
             FileWriter fw = new FileWriter(file);
             try {
-                fw.write(expectedCsv);
+                fw.write(s);
             } finally {
                 fw.close();
             }
